requirements.txt
flask
mistralai
python-dotenv
werkzeug